chore: rename docker plugin
